Use response error message if applicable
we send an error message from our end if something is wrong (or should),
so if it is an application level issue, we should show the error so that
it is easier for a user to report and see what is wrong.

we are also taking only the first error here, there could be more
technically but it should be fine for this case.